% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{preproc}
\alias{preproc}
\alias{preprocess}
\alias{preprocess.character}
\alias{preprocess.data.frame}
\alias{as_preprocessed}
\title{Preprocess}
\usage{
preprocess(text, id = NULL, count = TRUE, file = NULL)

\method{preprocess}{character}(text, id = NULL, count = TRUE,
  file = NULL)

\method{preprocess}{data.frame}(data, text, id = NULL, count = TRUE,
  file = NULL)

as_preprocessed(file)
}
\arguments{
\item{text}{Text to preprocess.}

\item{id}{Unique Document id.}

\item{count}{Whether to count term frequency.}

\item{file}{Path to \code{.txt} file where to save results. If omitted results are saved to a temp file.}

\item{data}{A \code{data.frame} containing \code{text} and (optionally) \code{id} columns.}
}
\description{
Convert text to \href{https://github.com/VowpalWabbit/vowpal_wabbit/wiki/Input-format}{Vowpal Wabbit} format.
}
